LUCIDDB/QA: Changed company tables for index/noindex to have same datatypes, more nightly ref file fixes, and removed junk test from company
[git-p4: depot-paths = "//open/lu/dev/": change = 6504]

+++ luciddb/test/sql/company/index.sql
@@ -31,13 +31,13 @@ create table DEPT (
 
 -- EMP: employees in the company
 create table EMP (
-  EMPNO integer primary key,
+  EMPNO numeric(5,0) primary key,
   FNAME varchar(20) not null,
   LNAME varchar(20) not null,
   SEX char(1),
   DEPTNO integer,
 -- references dept,
-  MANAGER integer,
+  MANAGER numeric(5,0),
   LOCID CHAR(2),
 -- references location,
   SAL integer,
